services/crypto_spot_monitor/requirements.txt
websockets==12.0
pyzmq==25.1.0
